// RescueBite/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  PROVIDER: 'PROVIDER',
  RECEIVER: 'RECEIVER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const UserStatus = {
  PENDING: 'PENDING',
  ACTIVE: 'ACTIVE',
  RESTRICTED: 'RESTRICTED',
  UNACTIVE: 'UNACTIVE',
  BLOCKED: 'BLOCKED',
  DELETED: 'DELETED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const AuthProvider = {
  GOOGLE: 'GOOGLE',
  CREDENTIAL: 'CREDENTIAL'
} as const

export type AuthProvider = (typeof AuthProvider)[keyof typeof AuthProvider]


export const BusinessType = {
  RESTAURANT: 'RESTAURANT',
  BAKERY: 'BAKERY',
  CAFE: 'CAFE',
  GROCERY: 'GROCERY',
  OTHER: 'OTHER'
} as const

export type BusinessType = (typeof BusinessType)[keyof typeof BusinessType]


export const FoodCategory = {
  BAKERY: 'BAKERY',
  GROCERY: 'GROCERY',
  PRODUCE: 'PRODUCE',
  DAIRY: 'DAIRY',
  MEAT: 'MEAT',
  PREPARED_MEAL: 'PREPARED_MEAL',
  BEVERAGE: 'BEVERAGE',
  OTHER: 'OTHER'
} as const

export type FoodCategory = (typeof FoodCategory)[keyof typeof FoodCategory]


export const FoodStatus = {
  DRAFT: 'DRAFT',
  AVAILABLE: 'AVAILABLE',
  PARTIALLY_RESERVED: 'PARTIALLY_RESERVED',
  FULLY_RESERVED: 'FULLY_RESERVED',
  EXPIRED: 'EXPIRED',
  CANCELLED: 'CANCELLED'
} as const

export type FoodStatus = (typeof FoodStatus)[keyof typeof FoodStatus]


export const ReservationStatus = {
  RESERVED: 'RESERVED',
  CANCELLED: 'CANCELLED',
  COMPLETED: 'COMPLETED',
  EXPIRED: 'EXPIRED'
} as const

export type ReservationStatus = (typeof ReservationStatus)[keyof typeof ReservationStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED',
  CANCELLED: 'CANCELLED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]